requirements.txt
firebase_admin==5.1.0
Flask==2.0.2
# pyrebase==3.0.27
Pyrebase4==4.5.0
requests==2.26.0
Werkzeug==2.0.2
